archive/adk_work/agents/requirements.txt
# Google ADK Dependencies
google-adk>=1.11.0
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
functions-framework>=3.4.0

# Database & Storage (Firebase/Firestore integration)
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0